Add in the concept of project logs and setup the templates to include them automatically if available
project logs are just a page within the projects category that has a type of "project-log" and a project_id of the projects ... id. They're ordered by date and display only the date and the contents. If a log entry is available, the projects index also links to the section