Building out a full gallery view page for photograps
based off of some inspiration from the guardian gallery lightboxes.

Still need to figure out how to do the next/prev links and I should
rework the photos to pull in a selection of the exif data (and maybe
allow for selectively including additional such as GEO for travel and
non hoome photos?) to list out on a table in the info box. And hook up
some light JS to show/hide the info box.